Add tags for proper collection of functions to tests
