+++ tests/test_argcheck.py
@@ -10,33 +10,33 @@ def test_argcheck(testfolder: str) -> None:
     benchmarks = os.path.join(testfolder, "benchmarks.py")
     r = runner.AbstractBenchmarkRunner()
     with pytest.raises(TypeError, match="expected type <class 'int'>.*"):
-        r.run(benchmarks, params={"x": 1, "y": "1"})
+        r.run(benchmarks, params={"x": 1, "y": "1"}, tags=("standard",))
     with pytest.raises(ValueError, match="missing value for required parameter.*"):
-        r.run(benchmarks, params={"x": 1})
+        r.run(benchmarks, params={"x": 1}, tags=("standard",))
 
-    r.run(benchmarks, params={"x": 1, "y": 1})
+    r.run(benchmarks, params={"x": 1, "y": 1}, tags=("standard",))
 
 
 def test_error_on_duplicate_params(testfolder: str) -> None:
-    benchmarks = os.path.join(testfolder, "duplicate_benchmarks.py")
+    benchmarks = os.path.join(testfolder, "benchmarks.py")
 
     with pytest.raises(TypeError, match="got incompatible types.*"):
         r = runner.AbstractBenchmarkRunner()
-        r.run(benchmarks, params={"x": 1, "y": 1})
+        r.run(benchmarks, params={"x": 1, "y": 1}, tags=("duplicate",))
 
 
 def test_log_warn_on_overwrite_default(
     testfolder: str, caplog: pytest.LogCaptureFixture
 ) -> None:
-    benchmark = os.path.join(testfolder, "default_benchmarks.py")
+    benchmark = os.path.join(testfolder, "benchmarks.py")
     r = runner.AbstractBenchmarkRunner()
     with caplog.at_level(logging.DEBUG):
-        r.run(benchmark, params={"a": 1})
+        r.run(benchmark, params={"a": 1}, tags=("with_default",))
     assert "using given value 1 over default value" in caplog.text
 
 
 def test_untyped_interface(testfolder: str) -> None:
-    benchmarks = os.path.join(testfolder, "untyped_benchmarks.py")
+    benchmarks = os.path.join(testfolder, "benchmarks.py")
 
     r = runner.AbstractBenchmarkRunner()
-    r.run(benchmarks, params={"value": 2})
+    r.run(benchmarks, params={"value": 2}, tags=("untyped",))
