+++ tests/test_runner.py
@@ -5,11 +5,13 @@ from nnbench.runner import AbstractBenchmarkRunner
 
 def test_runner_discovery(testfolder: str) -> None:
     r = AbstractBenchmarkRunner()
+    path = os.path.join(
+        testfolder, "contains_single_file_for_collection_test", "simple_benchmark.py")
 
-    r.collect(os.path.join(testfolder, "simple_benchmark.py"))
+    r.collect(path)
     assert len(r.benchmarks) == 1
 
     r.clear()
 
-    r.collect(testfolder)
+    r.collect(path)
     assert len(r.benchmarks) == 1
