+++ tests/testproject/contains_single_file_for_collection_test/simple_benchmark.py
