--- tests/testproject/default_benchmarks.py
@@ -1,6 +1,0 @@
-import nnbench
-
-
-@nnbench.benchmark
-def add_two(a: int = 0) -> int:
-    return a + 2
