--- tests/testproject/untyped_benchmarks.py
@@ -1,6 +1,0 @@
-import nnbench
-
-
-@nnbench.benchmark
-def increment(value):
-    return value + 1
